requirements.txt
beautifulsoup4
PyYAML
requests
plexapi
